requirements.txt
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.8.0
ament-cmake-test==1.3.11
ament-copyright==0.12.12
ament-cppcheck==0.12.12
ament-cpplint==0.12.12
ament-flake8==0.12.12
ament-index-python==1.4.0
ament-lint==0.12.12
ament-lint-cmake==0.12.12
ament-package==0.14.0
ament-pep257==0.12.12
ament-uncrustify==0.12.12
ament-xmllint==0.12.12
angles==1.15.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.11.0
bcrypt==4.0.1
beautifulsoup4==4.14.2
blinker==1.9.0
brotli==1.2.0
builtin-interfaces==1.2.1
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
cloudscraper==1.2.71
cobble==0.1.4
composition-interfaces==1.2.1
cryptography==46.0.3
cssselect2==0.8.0
cv-bridge==3.2.1
demo-nodes-py==0.20.5
diagnostic-msgs==4.8.0
dill==0.4.0
distlib==0.4.0
distro==1.9.0
domain-coordinator==0.10.0
EbookLib==0.18
ecdsa==0.19.1
einops==0.8.1
et_xmlfile==2.0.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.3
examples-rclpy-minimal-action-client==0.15.3
examples-rclpy-minimal-action-server==0.15.3
examples-rclpy-minimal-client==0.15.3
examples-rclpy-minimal-publisher==0.15.3
examples-rclpy-minimal-service==0.15.3
examples-rclpy-minimal-subscriber==0.15.3
exceptiongroup==1.3.1
exejs==0.0.6
fastapi==0.123.5
filelock==3.20.0
filetype==1.2.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.60.1
fsspec==2025.10.0
ftfy==6.3.1
geometry-msgs==4.8.0
google-auth==2.43.0
google-genai==1.52.0
greenlet==3.3.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
identify==2.6.15
idna==3.11
image-geometry==3.2.1
interactive-markers==2.3.2
itsdangerous==2.2.0
jh2==5.0.10
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
laser-geometry==2.4.0
launch==1.0.8
launch-ros==0.19.9
launch-testing==1.0.8
launch-testing-ros==0.19.9
launch-xml==1.0.8
launch-yaml==1.0.8
lifecycle-msgs==1.2.1
logging-demo==0.20.5
lxml==6.0.2
mammoth==1.11.0
map-msgs==2.1.0
markdown2==2.5.4
markdownify==1.2.2
marker-pdf==1.10.1
MarkupSafe==3.0.3
message-filters==4.3.7
mpmath==1.3.0
multiprocess==0.70.18
nav-msgs==4.8.0
networkx==3.4.2
niquests==3.15.2
nodeenv==1.9.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai==1.109.1
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
osrf-pycommon==2.1.6
packaging==25.0
passlib==1.7.4
pathos==0.3.4
pcl-msgs==1.0.0
pdftext==0.6.3
pendulum-msgs==0.20.5
pillow==10.4.0
platformdirs==4.5.0
pox==0.3.6
ppft==1.7.7
pre_commit==4.5.0
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydyf==0.11.0
PyMuPDF==1.26.6
PyMySQL==1.1.2
pyparsing==3.2.5
PyPDF2==3.0.1
pypdfium2==4.30.0
pyphen==0.17.2
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
python-pptx==1.0.2
python-qt-binding==1.1.2
PyYAML==6.0.3
qh3==1.5.6
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
RapidFuzz==3.14.3
rcl-interfaces==1.2.1
rclpy==3.3.16
rcutils==5.1.6
regex==2024.11.6
requests==2.32.5
requests-toolbelt==1.0.0
resource-retriever==3.1.3
rmw-dds-common==1.6.0
ros2action==0.18.12
ros2bag==0.15.14
ros2cli==0.18.12
ros2component==0.18.12
ros2doctor==0.18.12
ros2interface==0.18.12
ros2launch==0.19.9
ros2lifecycle==0.18.12
ros2multicast==0.18.12
ros2node==0.18.12
ros2param==0.18.12
ros2pkg==0.18.12
ros2run==0.18.12
ros2service==0.18.12
ros2topic==0.18.12
rosbag2-interfaces==0.15.14
rosbag2-py==0.15.14
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.6
rosidl-cli==3.1.6
rosidl-cmake==3.1.6
rosidl-generator-c==3.1.6
rosidl-generator-cpp==3.1.6
rosidl-generator-py==0.14.4
rosidl-parser==3.1.6
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.6
rosidl-typesupport-introspection-cpp==3.1.6
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sensor-msgs==4.8.0
sensor-msgs-py==4.8.0
shape-msgs==4.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.45
sros2==0.10.6
starlette==0.50.0
statistics-msgs==1.2.1
std-msgs==4.8.0
std-srvs==4.8.0
stereo-msgs==4.8.0
surya-ocr==0.17.0
sympy==1.14.0
teleop-twist-keyboard==2.4.0
tenacity==9.1.2
tf2-geometry-msgs==0.25.12
tf2-kdl==0.25.12
tf2-msgs==0.25.12
tf2-py==0.25.12
tf2-ros-py==0.25.12
tf2-tools==0.25.12
threadpoolctl==3.6.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tokenizers==0.22.1
topic-monitor==0.20.5
torch==2.9.1
tqdm==4.67.1
trajectory-msgs==4.8.0
transformers==4.57.3
translators==6.0.1
triton==3.5.1
turtlesim==1.4.2
typing-inspection==0.4.2
typing_extensions==4.15.0
unique-identifier-msgs==2.2.1
urllib3==2.5.0
urllib3-future==2.14.908
uvicorn==0.38.0
virtualenv==20.35.4
visualization-msgs==4.8.0
wassima==2.0.2
wcwidth==0.2.14
weasyprint==63.1
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.4
xlsxwriter==3.2.9
zopfli==0.4.0
